Refactored Auth code into relevant controllers and models
