+++ package.json
@@ -69,7 +69,6 @@
     "express-session": "^1.17.2",
     "cheerio": "^1.0.0-rc.10",
     "dotenv": "^10.0.0",
-    "eslint-config-airbnb": "^19.0.2",
     "express": "^4.17.1",
     "firebase": "^9.6.1",
     "openid-client": "^5.1.0",
